### 2022-09-01 #4727 @editorialbot
[PRE REVIEW]: IGraph/M: graph theory and network analysis for Mathematica
**Submitting author:** @szhorvat (<a href="http://orcid.org/0000-0002-3100-523X">Szabolcs Horvát</a>)
**Repository:** https://github.com/szhorvat/IGraphM
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.6.2
