### 2022-09-01 #4727 @szhorvat
Hm, I think editorialbot is looking at the branch which contains the paper (`joss-paper` branch). This is an orphan branch that only has the paper, but not the package. This might be why the line counts are invalid and it could not find a license. The relevant version of the package is on the `master` branch.

Do we need to place the paper on the same branch as the package? I would prefer to not mix them up, if not really necessary.
