### 2022-09-01 #4727 @szhorvat
> @szhorvat if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @)

Our suggestions are the following:

 - [Misha Lavrov](https://misha.fish/), not sure if he has a GitHub account, perhaps the editor can ask for it in email?
 - [Danial McDonald](https://resources.wolframcloud.com/publishers/resources?PublisherID=Daniel+McDonald), does not have a GitHub account, still possible to act as reviewer?
 - [yaroslavvb](https://community.wolfram.com/web/yaroslavvb)
 - [carlwoll](https://mathematica.stackexchange.com/search?q=user%3A45431+%5Bgraphs-and-networks%5D)
 - [jarikirma](https://mathematica.stackexchange.com/search?q=user%3A3056+%5Bgraphs-and-networks%5D)
 - [vsvor](https://arxiv.org/abs/2106.11824)
