### 2022-09-06 f02f10c
refactor: consolidate use of igMake / igMakeUnweighted
