+++ IGraphM/Isomorphism.m
@@ -3,7 +3,7 @@
 
 (* :Author: szhorvat *)
 (* :Date: 2018-10-24 *)
-(* :Copyright: (c) 2018-2020 Szabolcs Horvát *)
+(* :Copyright: (c) 2018-2022 Szabolcs Horvát *)
 
 Package["IGraphM`"]
 
@@ -91,7 +91,7 @@ IGGetIsomorphism::usage = "IGGetIsomorphism[graph1, graph2] gives one isomorphis
 SyntaxInformation[IGGetIsomorphism] = {"ArgumentsPattern" -> {_, _}};
 IGGetIsomorphism[graph1_?IGNullGraphQ, igraph2_?IGNullGraphQ] := {<||>}
 IGGetIsomorphism[graph1_?igGraphQ, graph2_?igGraphQ] :=
-    catch@Block[{ig1 = igMakeFast[graph1], ig2 = igMakeFast[graph2], result},
+    catch@Block[{ig1 = igMakeUnweighted[graph1], ig2 = igMakeUnweighted[graph2], result},
       result = igIndexVec@check@ig1@"getIsomorphism"[ManagedLibraryExpressionID[ig2]];
       If[result === {}, Return[{}]];
       List@AssociationThread[
@@ -107,7 +107,7 @@ IGGetSubisomorphism::usage = "IGGetSubisomorphism[subgraph, graph] gives one sub
 SyntaxInformation[IGGetSubisomorphism] = {"ArgumentsPattern" -> {_, _}};
 IGGetSubisomorphism[subgraph_?IGNullGraphQ, graph_?igGraphQ] := {<||>}
 IGGetSubisomorphism[subgraph_?igGraphQ, graph_?igGraphQ] :=
-    catch@Block[{ig1 = igMakeFast[graph], ig2 = igMakeFast[subgraph], result},
+    catch@Block[{ig1 = igMakeUnweighted[graph], ig2 = igMakeUnweighted[subgraph], result},
       result = igIndexVec@check@ig1@"getSubisomorphism"[ManagedLibraryExpressionID[ig2]];
       If[result === {}, Return[{}]];
       List@AssociationThread[
@@ -121,7 +121,7 @@ PackageExport["IGIsoclass"]
 IGIsoclass::usage = "IGIsoclass[graph] gives the isomorphism class of the graph. Used as the index into the vector returned by motif finding functions. See IGData to get lists of graphs ordered by isoclass.";
 
 SyntaxInformation[IGIsoclass] = {"ArgumentsPattern" -> {_}};
-IGIsoclass[graph_?igGraphQ] := Block[{ig = igMakeFast[graph]}, sck@ig@"isoclass"[]]
+IGIsoclass[graph_?igGraphQ] := Block[{ig = igMakeUnweighted[graph]}, sck@ig@"isoclass"[]]
 
 
 (* Helper functions for handling vertex and edge colours *)
@@ -220,7 +220,7 @@ amendUsage[IGBlissCanonicalLabeling,
 Options[IGBlissCanonicalLabeling] = { "SplittingHeuristics" -> "FirstLargest" };
 SyntaxInformation[IGBlissCanonicalLabeling] = {"ArgumentsPattern" -> {{__}, OptionsPattern[]}};
 IGBlissCanonicalLabeling[graph_?igGraphQ, opt : OptionsPattern[]] :=
-    catch@Block[{ig = igMakeFast[graph]},
+    catch@Block[{ig = igMakeUnweighted[graph]},
       blissCheckMulti[graph];
       AssociationThread[
         VertexList[graph],
@@ -228,7 +228,7 @@ IGBlissCanonicalLabeling[graph_?igGraphQ, opt : OptionsPattern[]] :=
       ]
     ]
[…]
