### 2022-09-06 13f70c0
refactor: get rid of last uses of igMakeFastWeighted
