### 2022-09-06 298c65f
cleanup: remove old commented code, last mentions of obsolete igMake variants
