### 2022-09-14 #4727 @kthyng
@szhorvat I appreciate your enthusiasm, but please leave it up to the JOSS editorial staff to reach out to editors and, in the future, reviewers. We do, however, appreciate the reviewer suggestions you gave above. This is a particularly busy time of year for everyone and we are all doing our best to engage in a timely manner.
