### 2022-09-27 443a77f
editor performance: introducing PreciseTracking
