+++ IGraphM/GraphEditor.m
@@ -11,6 +11,8 @@ Package["IGraphM`"]
 
 PackageExport["IGGraphEditor"]
 
+
+(*PDynamic = Dynamic*)
 (* Dev notes
  - foo[ Dynamic @ state_ ] is no better than HoldFirst etc. I just like this convention for GUI
  - state should not be modified during continuous actions
@@ -70,6 +72,7 @@ supportedGraphQ = ! MixedGraphQ[#] && SimpleGraphQ[#]&;
 (* ::Subsection::Closed:: *)
 (*iGraphEditor*)
 
+
 editorFailure[ msg_String ] := Failure[
   "GraphEditor"
 , <|"Message" -> "\[WarningSign] "<> msg |>
@@ -79,7 +82,6 @@ iGraphEditor[graph:((_? supportedGraphQ)|PatternSequence[]), opt:OptionsPattern[
 With[
   {
     packageFailure = editorFailure["Failed to load IGraphM`, make sure it is installed."]
-  , performanceFailure = editorFailure["Too many vertices and edges. Increase \"PerformanceLimit\" option to try anyway."]
   },
 Interpretation[
   {
@@ -87,35 +89,18 @@ Interpretation[
   , error = False
   , refresh
   }
-, refresh[] := Module[{temp}, Catch[
-    If[
-      Needs@"IGraphM`" === $Failed
-    , Throw[ error = packageFailure]
+, refresh[] := Module[{temp}
+  , Catch[
+      If[ Needs@"IGraphM`" === $Failed, Throw[ error = packageFailure]]
+
+    ; temp = geStateVersionCheck @ state
+    ; If[ AssociationQ @ temp, state = temp, Throw[ error = temp] ]
+
+    ; iGraphEditorInitialization[state, error]  (*can throw*)  
     ]
+  ]
 
-  ; temp = geStateVersionCheck @ state
-  ; If[ AssociationQ @ temp
-    , state = temp
-    , Throw[ error = temp]
-    ]
-
-  ; If[
-      state["config", "vCounter"] + state["config", "eCounter"] > state["config", "PerformanceLimit"]
-    , Throw[ error =  performanceFailure]
-    ]
-
-  ; error = False
-
-  ; geAction["UpdateVertexSize", Hold @ state]
-  ; geAction["UpdateEdgesShapes", Hold @ state]
-    (*(Hold) is there to workaround a bug with Interpretation's Initialization
-      which inserts evaluated Dynamic's arguments
-    *)
-
-  ]]
-
-;
-  PaneSelector[
+; PaneSelector[
   {
     True -> Button[Dynamic @ error,  refresh[], BaseStyle -> 15]
   , False -> Panel[
@@ -131,8 +116,33 @@ Interpretation[
 , GraphFromEditorState @ state
 
 , Initialization :> refresh[]
+, Deinitialization :> iGraphEditorDeinitialization[state]
 
 ]]
+
+
+iGraphEditorInitialization // Attributes = {HoldAll}
+iGraphEditorInitialization[state_, error_]:=Module[
+  {perfFailure = editorFailure["Too many vertices and edges. Increase \"PerformanceLimit\" option to try anyway."]}
+  
+, If[
+    state[ "vCounter"] + state[ "eCounter"] > state[ "PerformanceLimit"]
+  , Throw[ error =  perfFailure ]
+  ]
+
+; ToTrackedAssociation @ state
+  
+; geAction["UpdateVertexSize", Hold @ state]
+; geAction["UpdateEdgesShapes", Hold @ state]
+    (*(Hold) is there to workaround a bug with Interpretation's Initialization
[…]
