### 2022-09-29 be7f7df
editor performance: patch  mousedragged fired on mouseclick
