### 2022-10-03 #4727 @szhorvat
Hello @prashjha, my suggestions for reviewers are in this comment: https://github.com/openjournals/joss-reviews/issues/4727#issuecomment-1234689545 See towards the end of the comment for suggestions taken from the Google Sheets you linked.
