### 2022-10-04 7ca6b39
fix: mouseover performance regression
