### 2022-10-08 23940be
fix: local_density calculation for isolated nodes
