+++ IGraphM/LibraryResources/Source/nanoflann.hpp
@@ -70,6 +70,10 @@
 #undef min
 #endif
 #endif
+// Avoid conflicts with X11 headers
+#ifdef None
+#undef None
+#endif
 
 namespace nanoflann
 {
@@ -170,7 +174,7 @@ class KNNResultSet
     CountType     count;
 
    public:
-    inline KNNResultSet(CountType capacity_)
+    explicit inline KNNResultSet(CountType capacity_)
         : indices(0), dists(0), capacity(capacity_), count(0)
     {
     }
@@ -257,7 +261,7 @@ class RadiusResultSet
 
     std::vector<std::pair<IndexType, DistanceType>>& m_indices_dists;
 
-    inline RadiusResultSet(
+    explicit inline RadiusResultSet(
         DistanceType                                     radius_,
         std::vector<std::pair<IndexType, DistanceType>>& indices_dists)
         : radius(radius_), m_indices_dists(indices_dists)
@@ -394,7 +398,9 @@ struct L1_Adaptor
         /* Process last 0-3 components.  Not needed for standard vector lengths.
          */
         while (a < last)
-        { result += std::abs(*a++ - data_source.kdtree_get_pt(b_idx, d++)); }
+        {
+            result += std::abs(*a++ - data_source.kdtree_get_pt(b_idx, d++));
+        }
         return result;
     }
 
@@ -650,25 +656,31 @@ struct metric_SO3 : public Metric
 /** @addtogroup param_grp Parameter structs
  * @{ */
 
-enum class KDTreeSingleIndexAdaptorFlags {
-    None = 0,
+enum class KDTreeSingleIndexAdaptorFlags
+{
+    None                  = 0,
     SkipInitialBuildIndex = 1
 };
 
-inline std::underlying_type<KDTreeSingleIndexAdaptorFlags>::type operator&(KDTreeSingleIndexAdaptorFlags lhs, KDTreeSingleIndexAdaptorFlags rhs) {
-    using underlying = typename std::underlying_type<KDTreeSingleIndexAdaptorFlags>::type;
+inline std::underlying_type<KDTreeSingleIndexAdaptorFlags>::type operator&(
+    KDTreeSingleIndexAdaptorFlags lhs, KDTreeSingleIndexAdaptorFlags rhs)
+{
+    using underlying =
+        typename std::underlying_type<KDTreeSingleIndexAdaptorFlags>::type;
     return static_cast<underlying>(lhs) & static_cast<underlying>(rhs);
 }
 
 /**  Parameters (see README.md) */
 struct KDTreeSingleIndexAdaptorParams
 {
-    KDTreeSingleIndexAdaptorParams(size_t _leaf_max_size = 10, KDTreeSingleIndexAdaptorFlags _flags = KDTreeSingleIndexAdaptorFlags::None)
+    KDTreeSingleIndexAdaptorParams(
+        size_t _leaf_max_size = 10, KDTreeSingleIndexAdaptorFlags _flags =
+                                        KDTreeSingleIndexAdaptorFlags::None)
         : leaf_max_size(_leaf_max_size), flags(_flags)
     {
     }
 
-    size_t leaf_max_size;
+    size_t                        leaf_max_size;
     KDTreeSingleIndexAdaptorFlags flags;
 };
 
@@ -871,7 +883,8 @@ struct array_or_vector_selector<-1, T>
  * and KDTreeSingleIndexDynamicAdaptor_.
  *
  * \tparam Derived The name of the class which inherits this class.
- * \tparam DatasetAdaptor The user-provided adaptor (see comments above).
+ * \tparam DatasetAdaptor The user-provided adaptor, which must be ensured to
+ *         have a lifetime equal or longer than the instance of this class.
[…]
