### 2022-10-08 6cf7271
fix: work around mishandling of lists by VertexIndex in M12.0+, fixes #121
