### 2022-10-08 8861250
fix: eliminate most usages of VertexIndex due to the quirks of this function in M12.0+
