### 2022-10-11 4cbcd81
feat: experimental local density calculation supports non-simple graphs
