### 2022-10-11 #4727 @szhorvat
@prashjha A week has passed since inviting chaypaterson. He has almost zero activity on GitHub, and has not done any reviews for JOSS before, so it seems reasonable to assume that he did not get the note. I am wondering if it's time to attempt email contact.
