### 2022-10-12 #4727 @kthyng
@szhorvat Please understand that everyone here, from the editors through the reviewers, are volunteers with limited time, just like you. Unless you are worried about a violation of journal policies or something that you need to bring to an editor's attention, please let us all do our jobs without your prodding. It is inappropriate for the author to attempt to direct the review process.
