### 2022-10-12 #4727 @szhorvat
Dear @kthyng,

I mean to cause no pressure. However, considering that the submission was made over 6 weeks ago, and reviewers have not been selected yet, the process appears to be considerably slower than what I expected based on my experience with other journals, either in the role of author or reviewer. Unfortunately, the delay does affect me, and not knowing what to expect, I am becoming concerned. At this point I would have enquired about progress with any other journal through their appropriate contact channels, so please consider this message that enquiry. It was certainly not meant as an attempt to direct the review process. Given the differences between JOSS and more traditional journals (including the open nature of article processing), I believe authors would find it helpful to receive some feedback about expected processing times.
