### 2022-10-12 #4727 @arfon
> I mean to cause no pressure.

But your actions *are* causing pressure @szhorvat, and as JOSS EiC that is a concern to me. I will note that twice already during the pre-review of this submission, you have chased the editorial team for an update.

JOSS operates an open review system, which means there is no priviliged information. That has its advantages (everyone knows exactly what state the submission is in) and disadvantages (everyone knows exactly what state the submission is in) which include authors sometimes chasing editors for an update.

Going forward, I would ask you to not request an update from the editor unless it has been at least two weeks since a last response.
