+++ Tests/Tests.m
@@ -5250,6 +5250,32 @@ MT[
 ]
 
 
+(* ::Subsection::Closed:: *)
+(*Modularity*)
+
+
+MT[
+  IGModularity[Graph[{1 <-> 2, 2 <-> 3, 3 <-> 4, 4 <-> 1, 1 <-> 2, 3 <-> 3}], {{2, 3}, {1, 4}}],
+  -0.013888888888888784
+]
+
+MT[
+  IGModularityMatrix[Graph[{1 <-> 2, 2 <-> 3, 3 <-> 4, 4 <-> 1, 1 <-> 2, 3 <-> 3}]],
+  {{-0.75, 1.25, -1., 0.5}, {1.25, -0.75, 0., -0.5}, {-1., 0., 0.6666666666666667, 0.33333333333333337}, {0.5, -0.5, 0.33333333333333337, -0.3333333333333333}}
+]
+
+
+MT[
+  IGModularity[CycleGraph[4], {{2, 3}, {1, 4}}],
+  0.
+]
+
+MT[
+  IGModularityMatrix[CycleGraph[4]],
+  {{-0.5, 0.5, -0.5, 0.5}, {0.5, -0.5, 0.5, -0.5}, {-0.5, 0.5, -0.5, 0.5}, {0.5, -0.5, 0.5, -0.5}}
+]
+
+
 (* ::Section::Closed:: *)
 (*Shortest paths*)
 
