### 2022-10-22 #4727 @prashjha
Hi @chaypaterson, I am wondering if you saw my previous ping. I think this JOSS submission could be of interest to you. Please let me know if you can review this submission. Thanks!!
