### 2022-10-22 #4727 @chaypaterson
Hi Prashant

Sorry, I am afraid I am unavailable to review. I will not have the time.

Chay

On Sat, 22 Oct 2022, 04:14 Prashant K. Jha, ***@***.***>
wrote:

> Hi @chaypaterson <https://github.com/chaypaterson>, I am wondering if you
> saw my previous ping. I think this JOSS submission could be of interest to
> you. Please let me know if you can review this submission. Thanks!!
>
