### 2022-10-28 18194e6
editor: multi edges and self loops
