+++ IGraphM/GraphEditor.m
@@ -66,7 +66,7 @@ IGGraphEditor /:
 iGraphEditor // Options = Options @ IGGraphEditor;
 
 
-supportedGraphQ = ! MixedGraphQ[#] && SimpleGraphQ[#]&;
+supportedGraphQ = MatchQ[_Graph];
 
 
 (* ::Subsection:: *)
@@ -340,7 +340,7 @@ stateVertexList[state_Association] := Values @ state[["vertex", All, "name"]]
 
 
 stateEdgeList[state_Association] := state //
- Query["edge", Values, ((#type /. # /. state[["vertex", All, "name"]])) &]
+ Query["edge", Values, (#edge /. state[["vertex", All, "name"]]) &]
 
 
 stateGraphEmbedding[state_Association] := state // Query["vertex", Values, "pos"]
@@ -491,20 +491,23 @@ If[ $VersionNumber > 12
 ]
 
 
-edgeToPrimitive[e_] := If[
-  e["shape"] =!= Automatic
-, e["shape"]
-, edgeTypeToPrimitive[e["type"]
-][
-    DynamicLocation[e["v1"], Automatic],
-    DynamicLocation[e["v2"], Automatic]
+edgeToPrimitive[e_] := Module[{shapeFunction}
+
+, shapeFunction = If[ e["shape"] =!= Automatic
+  , Return[ e["shape"], Module ]
   ]
-]
+
+; edgeTypeToPrimitive[ e["edge"] ][
+    DynamicLocation[e[["edge", 1]], Automatic],
+    DynamicLocation[e[["edge", 2]], Automatic]
+  ]
+]  
 
 
-edgeTypeToPrimitive["v1"->"v2"] = Arrow[{#,#2}]&
-edgeTypeToPrimitive["v2"->"v1"] = Arrow[{#2,#}]&
-edgeTypeToPrimitive[UndirectedEdge["v1", "v2"]] = Line[{#,#2}]&
+edgeTypeToPrimitive[(Rule|DirectedEdge)[_,_]]=Arrow[{#,#2}]&
+edgeTypeToPrimitive[_[_, _]]=Line[{#,#2}]&
+
+
 
 
 (* ::Subsubsection::Closed:: *)
@@ -777,7 +780,7 @@ smallestMissingInteger[list_List] := Block[{n = 1}
 
 geAction["RemoveVertex", Dynamic @ state_, v_] := With[{ id = v["id"]}
 , state["vertex"]  = KeyDrop[id] @ state["vertex"]
-; state["edge"]    = DeleteCases[state["edge"], KeyValuePattern[{_ -> id}] ]
+; state["edge"]    = Select[state["edge"], FreeQ[#edge, id] & ]
 ; state["vCounter"]--
 ; state["eCounter"] = Length @ state["edge"]
 ; If[ state["selectedVertex"] === id, geAction["Unselect", Dynamic @ state] ]
@@ -789,15 +792,9 @@ geAction["RemoveVertex", Dynamic @ state_, v_] := With[{ id = v["id"]}
 
 
 geAction["CreateEdge", Dynamic @ state_, selectedV_String, clickedV_String] := Module[{eId, type}
-, If[
-    Not @ newEdgeAllowedQ[state, selectedV, clickedV]
-  , Beep[]
-  ; Message[IGGraphEditor::multiEdge]
-  ; Return[$Failed, Module]
-  ]
-; eId = CreateUUID["e-"]
+, eId = CreateUUID["e-"]
 
-; type =   If[state["DirectedEdges"], Rule, UndirectedEdge]
+; type = If[state["DirectedEdges"], Rule, UndirectedEdge]
 
 ; state["edge", eId ] = createEdge[eId,  type[selectedV, clickedV] ]
 ; state["eCounter"]++
[…]
