### 2022-10-28 a0630df
editor fix: sync state after vertex drag
