### 2022-11-02 #4727 @pitsianis
I will be happy to help. I have used iGraph through Leiden and a Leiden wrapper to access it from Julia.

@arfon what is the JOSS policy, if any, regarding software wrappers that make a binary library accessible from a different language?
