### 2022-11-02 20afe63
editor: creating self loops
