+++ IGraphM/GraphEditor.m
@@ -36,6 +36,89 @@ $potentialEdgeStyle = Directive[Purple, Dashed, Thick];
 $gridLinesCount = 30.;
 
 
+(* ::Subsection::Closed:: *)
+(*Helpers*)
+
+
+es6Decorate // Attributes={HoldAll};
+
+es6Decorate[f_Symbol]:=(
+
+  f /: SetDelayed[
+    f[ Verbatim[Association][spec__] ]
+    , rhs_
+  ]:=With[
+    { pattern    = es6ExtractPattern[spec]
+      , heldSymbols = es6ExtractSymbols[spec]
+      , heldValues  = es6ExtractValues[spec]
+    }
+    , es6Decorate[f, rhs, pattern, heldSymbols, heldValues]
+  ]
+);
+
+es6Decorate[foo_Symbol,rhs_,pattern_, Hold[symbols___], Hold[values___]]:=(
+  SetDelayed @@ Hold[
+    foo[$ES6asso:pattern],
+    Block[{symbols}, Unevaluated[rhs] /. Thread[{symbols}->{values}] ]
+  ]
+)
+
+es6ExtractPattern[spec__]:= KeyValuePattern[
+  es6PatternToPattern /@ {spec}
+]
+
+es6ExtractSymbols[spec__]:= Apply[Join][
+  es6PatternToHeldSymbol /@ {spec}
+]
+
+es6ExtractValues[spec__]:= Apply[Join][
+  es6PatternToHeldValue /@ {spec}
+]
+
+With[
+  { vPatt = Verbatim[Pattern]
+    , vOpt  = Verbatim[Optional]
+    , hPatt = HoldPattern
+  }
+  ,
+
+  (* a_ a_H *)
+  es6PatternToPattern[    hPatt @ vPatt[s_Symbol, b_Blank] ] := symbolName[s] -> b;
+  (* "a" \[Rule] b_ *)
+  es6PatternToPattern[    sym_String -> Except[_Optional] ]:= sym -> Blank[];
+  es6PatternToPattern[ ___ ] = Sequence[];
+
+
+
+  es6PatternToHeldSymbol[ hPatt @ vPatt[s_Symbol, _]      ] := Hold[s];
+  es6PatternToHeldSymbol[ hPatt @ vOpt[p_, _]      ] := es6PatternToHeldSymbol @ p;
+  es6PatternToHeldSymbol[ _String -> p_ ]:= es6PatternToHeldSymbol @ p;
+  es6PatternToHeldSymbol[ ___ ]:=Hold[];
+
+
+
+  (* s_ ==> ass["s"] *)
+  es6PatternToHeldValue[  hPatt @ vPatt[s_Symbol, _]      ] := With[{sym = symbolName@s}
+    , Hold @ $ES6asso @ sym
+  ];
+  (* s_:default ==> Lookup[asso, "s", default] *)
+  es6PatternToHeldValue[  hPatt @ vOpt[vPatt[s_Symbol, _Blank], default_] ] := With[{sym = symbolName@s}
+    , Hold @ Lookup[$ES6asso, sym, default]
+  ];
+  (* S \[Rule] s_ *)
+  es6PatternToHeldValue[ key_ -> hPatt @ vPatt[_, _Blank] ]:= Hold @ $ES6asso @ key;
+
+  (* S \[Rule] s_:default *)
+  es6PatternToHeldValue[ key_ -> hPatt @ vOpt[vPatt[_, _Blank], default_] ]:=
+    Hold @ Lookup[$ES6asso, key, default];
+  es6PatternToHeldValue[  ___     ] := Hold[];
+];
+
[…]
