### 2022-11-02 20af626
fix: replace deprecated igraph_is_inf() by isinf()
