### 2022-11-03 #4899 @leotrs
Hello @szhorvat just noting here that I tried to run your software on the [Wolfram Cloud](https://www.wolframcloud.com/) and it won't run. The installer prints a nice error message in this case, which is great! So thanks for that. The reason seems to be that a necessary dependency called `LibraryLink` is not supported on Wolfram Cloud.

Is `LibraryLink` 100% needed? If your software was able to run on Wolfram Cloud, which doesn't require buying a license from Wolfram Research, it would greatly increase the audience of your package. If this is completely necessary, then I understand, I just thought I'd ask.

As an aside, my review will be delayed a couple of days as I get my hands on a Mathematica license so I can test the software.
