### 2022-11-03 #4899 @szhorvat
The bulk of the library is written in C and C++, which allows it to have good performance, and which is necessary for interfacing with several dependencies. Wolfram Cloud disables not just LibraryLink, but _all_ means of running arbitrary C code (I assume for security reasons), so unfortunately there is no way to make the package compatible with the Wolfram Cloud. I am hoping that Wolfram will change this at some point in the future.

To get this out of the way:

There is the ["Free Wolfram Engine"](https://www.wolfram.com/engine/), which you can install on your own machine. IGraph/M _does_ work with this. However, the Free Wolfram Engine does not include the graphical notebook interface. I think this changes to user experience to such a degree that a proper review is not possible: It won't be possible to run documentation examples directly, there won't be auto-completion (the package has long function names, so this is essential) and none of the `Dynamic` functionality will be accessible. Therefore I would like to ask reviewers not to base their assessment purely on working with the Free Wolfram Engine either.
