### 2022-11-03 #4899 @leotrs
This makes perfect sense, thank you! I would suggest that you add a note regarding these matters to your package's README. Specifically, I'd suggest mentioning that your package does not work in the Wolfram Cloud, and that the Wolfram Engine works but is not recommended.
