### 2022-11-06 #4899 @szhorvat
I added a note in the readme about the Wolfram Cloud.

As for the Free Wolfram Engine, my comment above referred specifically to reviewing. Using the Free Wolfram Engine necessitates using the command line interface instead of the notebook interface. This makes for a frustrating experience, and does not represent how the vast majority of people use Mathematica. Therefore, it should not be the _sole_ basis of reviews. This comment would apply equally to any other package, not just IGraph/M.

The same applies to using Jupyter with the Free Wolfram Engine, see https://mathematica.stackexchange.com/q/274333/12

That said, I do make sure that IGraph/M works well with the Free Wolfram Engine, and do test it occasionally. Command line usage does have its place in special circumstances (e.g. I use IGraph/M that way on our HPC cluster). You are welcome to try it out.
