### 2022-11-06 #4899 @szhorvat
For the purpose of organizing the reviews better, you are welcome to [open issues in the issue tracker](https://github.com/szhorvat/IGraphM/issues/new/choose) even if they are not about a bug or feature request. The advantage compared to the forum (which we normally recommend) is that you can put a backlink to this issue.
