### 2022-11-09 #4899 @leotrs
I was able to install and use it with ease. I have a soft spot for Mathematica AND graph theory, so I love this one. Thanks @szhorvat. I believe the paper can be accepted as is.
