### 2022-11-22 c768b74
feat: low-level integration for igraph_triangular_lattice
