+++ IGraphM/IGraphM.m
@@ -206,6 +206,7 @@ template = LTemplate["IGraphM",
         LFun["realizeDegreeSequence", {{Integer, 1, "Constant"} (* outdeg *), {Integer, 1, "Constant"} (* indeg *), True|False (* loops *), True|False (* multi *), Integer (* method *)}, "Void"],
         LFun["fromLCF", {Integer, {Integer, 1, "Constant"}, Integer}, "Void"],
         LFun["makeLattice", {{Integer, 1, "Constant"}, Integer (* nei *), True|False (* directed *), True|False (* mutual *), True|False (* periodic *)}, "Void"],
+        LFun["makeTriLattice", {{Integer, 1, "Constant"}, True|False (* directed *), True|False (* mutual *)}, "Void"],
         LFun["kautz", {Integer, Integer}, "Void"],
         LFun["tree", {Integer, Integer, True|False (* directed *)}, "Void"],
         LFun["fromPrufer", {{Integer, 1, "Constant"}}, "Void"],
