### 2022-11-22 81170ad
fix: remove superfluous newline from usage
