### 2022-11-22 f8c8ba0
perf: performance improvement for proximity graph calculations
