+++ IGraphM/LibraryResources/Source/IGFlann3D.h
@@ -8,8 +8,8 @@
 #define IG_FLANN3D_H
 
 #include "nanoflann.hpp"
+#include "IGCommon.h"
 #include "IGFlannCommon.h"
-#include <LTemplate.h>
 #include <vector>
 
 using namespace nanoflann;
@@ -31,6 +31,91 @@ class IGFlann3D {
         delete kdtree;
         delete ps;
     }
+
+
+    class NeighborCounts {
+        const double radius; // L2 search radius
+        const bool short_circuit; // whether to stop after one point has been found
+        const mint v1, v2; // edge endpoits; excluded from the count
+        const PointSet *ps;
+
+        mint count;
+
+    public:
+        NeighborCounts(
+                    double radius_, bool short_circuit_,
+                    mint v1_, mint v2_, const PointSet *ps_)
+            : radius(radius_), short_circuit(short_circuit_), v1(v1_), v2(v2_), ps(ps_)
+        {
+            init();
+        }
+
+        void init() { clear(); }
+        void clear() { count = 0; }
+
+        size_t size() const { return count; }
+
+        bool full() const { return true; }
+
+        bool addPoint(double dist, mint index) {
+            if (dist < radius && index != v1 && index != v2) {
+                count++;
+                if (short_circuit)
+                    return false;
+            }
+            return true;
+        }
+
+        double worstDist() const { return radius; }
+    };
+
+
+    class IntersectionCounts {
+        const double radius; // L2 search radius
+        const double beta_radius;
+        const bool short_circuit; // whether to stop after one point has been found
+        const mint v1, v2; // edge endpoits; excluded from the count
+        const double *centre1, *centre2; // disc centres
+        const PointSet *ps;
+
+        size_t count;
+
+    public:
+        IntersectionCounts(
+                    double radius_, double beta_radius_,
+                    bool short_circuit_,
+                    mint v1_, mint v2_, const double *centre1_, const double *centre2_, const PointSet *ps_)
+            : radius(radius_), beta_radius(beta_radius_),
+              short_circuit(short_circuit_),
+              v1(v1_), v2(v2_),
+              centre1(centre1_), centre2(centre2_), ps(ps_)
+        {
+            init();
+        }
+
+        void init() { clear(); }
+        void clear() { count = 0; }
+
+        size_t size() const { return count; }
+
+        bool full() const { return true; }
+
+        bool addPoint(double dist, mint index) {
+            if (dist < radius && index != v1 && index != v2) {
+                double pd1 = sqdist3(&ps->pts(index,0), centre1);
+                double pd2 = sqdist3(&ps->pts(index,0), centre2);
+                if (pd1 < beta_radius && pd2 < beta_radius) {
+                    count++;
+                    if (short_circuit)
+                        return false;
+                }
+            }
+            return true;
+        }
+
+        double worstDist() const { return radius; }
+    };
+
 
 public:
 
@@ -107,10 +192,13 @@ public:
 
 
     // how many neighbours does each point have within the given distance?
+    // if short_circuit == true then 1 will be returned whenever at least one point exists
+    // edge endpoints are excluded
     mma::IntTensorRef neighborCounts(
             mma::RealMatrixRef pts,
             mma::RealTensorRef dists,
-            mma::IntMatrixRef edges) {
+            mma::IntMatrixRef edges,
+            bool short_circuit) {
 
         mint n = dists.size();
 
@@ -127,34 +215,31 @@ public:
         std::vector<std::pair<mint, double>> ret_matches;
 
         auto res = mma::makeVector<mint>(n);
+        LTGuard<mma::IntTensorRef> res_guard(res);
 
[…]
