+++ IGraphM/IGraphM.m
@@ -670,19 +670,22 @@ template = LTemplate["IGraphM",
         LFun["neighborCounts",
           {{Real, 2, "Constant"} (* centres *),
            {Real, 1, "Constant"} (* distances *),
-           {Integer, 2, "Constant"} (* edges *)},
+           {Integer, 2, "Constant"} (* edges *),
+           True|False (* short_circuit *)},
           {Integer, 1}],
         LFun["intersectionCounts",
           {{Real, 2, "Constant"} (* centres1 *),
            {Real, 2, "Constant"} (* centres2 *),
            {Real, 1, "Constant"} (* distances *),
-           {Integer, 2, "Constant"} (* edges *)},
+           {Integer, 2, "Constant"} (* edges *),
+           True|False (* short_circuit *)},
           {Integer, 1}],
         LFun["unionCounts",
           {{Real, 2, "Constant"} (* centres1 *),
            {Real, 2, "Constant"} (* centres2 *),
            {Real, 1, "Constant"} (* distances *),
-           {Integer, 2, "Constant"} (* edges *)},
+           {Integer, 2, "Constant"} (* edges *),
+           True|False (* short_circuit *)},
           {Integer, 1}]
       }
     ],
@@ -695,19 +698,22 @@ template = LTemplate["IGraphM",
         LFun["neighborCounts",
           {{Real, 2, "Constant"} (* centres *),
            {Real, 1, "Constant"} (* distances *),
-           {Integer, 2, "Constant"} (* edges *)},
+           {Integer, 2, "Constant"} (* edges *),
+           True|False (* short_circuit *)},
           {Integer, 1}],
         LFun["intersectionCounts",
           {{Real, 2, "Constant"} (* centres1 *),
            {Real, 2, "Constant"} (* centres2 *),
            {Real, 1, "Constant"} (* distances *),
-           {Integer, 2, "Constant"} (* edges *)},
+           {Integer, 2, "Constant"} (* edges *),
+           True|False (* short_circuit *)},
           {Integer, 1}],
         LFun["unionCounts",
           {{Real, 2, "Constant"} (* centres1 *),
            {Real, 2, "Constant"} (* centres2 *),
            {Real, 1, "Constant"} (* distances *),
-           {Integer, 2, "Constant"} (* edges *)},
+           {Integer, 2, "Constant"} (* edges *),
+           True|False (* short_circuit *)},
           {Integer, 1}]
       }
     ]
