### 2022-11-22 46d56a6
perf: performance improvement for proximity graph calculations
