+++ IGraphM/LibraryResources/Source/IGFlann2D.h
@@ -9,7 +9,7 @@
 
 #include "nanoflann.hpp"
 #include "IGFlannCommon.h"
-#include <LTemplate.h>
+#include "IGCommon.h"
 #include <vector>
 
 using namespace nanoflann;
@@ -31,6 +31,95 @@ class IGFlann2D {
         delete kdtree;
         delete ps;
     }
+
+
+    // Helper result class to count points within a neighbourhood of a given radius.
+    // Set short_circuit = true to quickly determine whether the neighbourhood
+    // contains at least one point. In this case size() returns 0 if no point
+    // was found, and 1 if at least one point was found.
+    class NeighborCounts {
+        const double radius; // L2 search radius
+        const bool short_circuit; // whether to stop after one point has been found
+        const mint v1, v2; // edge endpoits; excluded from the count
+        const PointSet *ps;
+
+        size_t count;
+
+    public:
+        NeighborCounts(double radius_, bool short_circuit_,
+                       mint v1_, mint v2_, const PointSet *ps_)
+            : radius(radius_), short_circuit(short_circuit_), v1(v1_), v2(v2_), ps(ps_)
+        {
+            init();
+        }
+
+        void init() { clear(); }
+        void clear() { count = 0; }
+
+        size_t size() const { return count; }
+
+        bool full() const { return true; }
+
+        bool addPoint(double dist, mint index) {
+            if (dist < radius && index != v1 && index != v2) {
+                count++;
+                if (short_circuit)
+                    return false;
+            }
+            return true;
+        }
+
+        double worstDist() const { return radius; }
+    };
+
+
+    // Helper result class for listing the points in the lune's bounding disk
+    // and counting how many are within the lune.
+    class IntersectionCounts {
+        const double radius; // L2 search radius
+        const double beta_radius;
+        const bool short_circuit; // whether to stop after one point has been found
+        const mint v1, v2; // edge endpoits; excluded from the count
+        const double *centre1, *centre2; // disc centres
+        const PointSet *ps;
+
+        size_t count;
+
+    public:
+        IntersectionCounts(
+                    double radius_, double beta_radius_,
+                    bool short_circuit_,
+                    mint v1_, mint v2_, const double *centre1_, const double *centre2_, const PointSet *ps_)
+            : radius(radius_), beta_radius(beta_radius_),
+              short_circuit(short_circuit_),
+              v1(v1_), v2(v2_),
+              centre1(centre1_), centre2(centre2_), ps(ps_)
+        {
+            init();
+        }
+
+        void init() { clear(); }
+        void clear() { count = 0; }
+
+        size_t size() const { return count; }
+
+        bool full() const { return true; }
+
+        bool addPoint(double dist, mint index) {
+            if (dist < radius && index != v1 && index != v2) {
+                double pd1 = sqdist2(&ps->pts(index,0), centre1);
+                double pd2 = sqdist2(&ps->pts(index,0), centre2);
+                if (pd1 < beta_radius && pd2 < beta_radius) {
+                    count++;
+                    if (short_circuit)
+                        return false;
+                }
+            }
+            return true;
+        }
+
+        double worstDist() const { return radius; }
+    };
 
 public:
 
@@ -107,10 +196,13 @@ public:
 
 
     // how many neighbours does each point have within the given distance?
+    // if short_circuit == true then 1 will be returned whenever at least one point exists
+    // edge endpoints are excluded
     mma::IntTensorRef neighborCounts(
             mma::RealMatrixRef pts,
             mma::RealTensorRef dists,
[…]

[…]
