### 2022-11-24 d6f7984
feat: IGBetaWeightedGabrielGraph
