+++ IGraphM/IGraphM.m
@@ -686,7 +686,11 @@ template = LTemplate["IGraphM",
            {Real, 1, "Constant"} (* distances *),
            {Integer, 2, "Constant"} (* edges *),
            True|False (* short_circuit *)},
-          {Integer, 1}]
+          {Integer, 1}],
+        LFun["edgeBetas",
+          {{Integer, 2, "Constant"} (* edges *),
+           Real (* maxBeta *)},
+          {Real, 1}]
       }
     ],
 
@@ -714,7 +718,11 @@ template = LTemplate["IGraphM",
            {Real, 1, "Constant"} (* distances *),
            {Integer, 2, "Constant"} (* edges *),
            True|False (* short_circuit *)},
-          {Integer, 1}]
+          {Integer, 1}],
+        LFun["edgeBetas",
+          {{Integer, 2, "Constant"} (* edges *),
+           Real (* maxBeta *)},
+          {Real, 1}]
       }
     ]
   }
