### 2022-11-28 f234edc
IGBetaWeightedGabrielGraph: add support for tolerances
