### 2022-12-06 #4899 @szhorvat
I noticed that `@pitsianis` has checked all [the review checkboxes](https://github.com/openjournals/joss-reviews/issues/4899#issuecomment-1299649682). Thank you for taking a look at the package! However, since there was no verbal feedback, I would like to clarify whether this constitutes a completion of the review.

Due to recent changes in my circumstances, I need need to freeze work on this project for a while starting early next year. If at all possible, it would be very helpful for me to have an estimate of the review timeline, so I can plan accordingly.
