### 2022-12-06 #4899 @pitsianis
I believe the reviewers are done. Great work. I implemented a wrapper of Leiden for Julia recently. Leiden makes good use of IGraph.
