### 2022-12-17 8d93fbe
flesh out IGLuneWeightedGabrielGraph
