### 2022-12-17 696932a
paper: minor typo, update for 0.6.4
