+++ paper.md
@@ -63,7 +63,7 @@ The only other independent graph theory package for Mathematica was [Combinatori
 
 # Design goals and features
 
-One of the major appeals of Mathematica is its tightly integrated nature: different functionality areas of the system can smoothly and seamlessly interoperate with each other. In order to preserve this benefit, a major design goal for IGraph/M was to integrate well into the rest of the system. This is achieved by working directly with Mathematica's native `Graph` data type, which is transparently converted to igraph's internal format as needed. This makes IGraph/M different from igraph's other high-level interfaces: igraph's internal graph data structure is not exposed to users and the package does not provide trivial operations which are already present in Mathematica, such as adding or removing vertices and edges. Instead, priority is given to functionality that delivers a true benefit over Mathematica's own built-ins. While some of IGraph/M's functions may appear to simply duplicate built-in functions, almost all of them provide additional features not otherwise available in Mathematica. For example, unlike the built-in `BetweenessCentrality[]` function, `IGBetweenness[]` supports weighted graphs; unlike the built-in `PageRankCentrality[]`, `IGPageRank[]` takes into consideration self-loops and parallel edges; in contrast to `IsomorphicGraphQ`, `IGIsomorphicQ` supports non-simple graphs as well as vertex and edge colours; etc. In order to make it easy to distinguish built-in symbols from those of IGraph/M, all functions in this package have names starting with `IG`. The naming and interface of functions is chosen so as to be familiar both to Mathematica users and users of igraph's Python and R interfaces.
+One of the major appeals of Mathematica is its tightly integrated nature: different functionality areas of the system can smoothly and seamlessly interoperate with each other. In order to preserve this benefit, a major design goal for IGraph/M was to integrate well into the rest of the system. This is achieved by working directly with Mathematica's native `Graph` data type, which is transparently converted to igraph's internal format as needed. This makes IGraph/M different from igraph's other high-level interfaces: igraph's internal graph data structure is not exposed to users and the package does not provide trivial operations which are already present in Mathematica, such as adding or removing vertices and edges. Instead, priority is given to functionality that delivers a true benefit over Mathematica's own built-ins. While some of IGraph/M's functions may appear to simply duplicate built-in functions, almost all of them provide additional features not otherwise available in Mathematica. For example, unlike the built-in `BetweenessCentrality[]` function, `IGBetweenness[]` supports weighted graphs; unlike the built-in `PageRankCentrality[]`, `IGPageRank[]` takes into consideration self-loops and parallel edges; in contrast to `IsomorphicGraphQ[]`, `IGIsomorphicQ[]` supports non-simple graphs as well as vertex and edge colours; etc. In order to make it easy to distinguish built-in symbols from those of IGraph/M, all functions in this package have names starting with `IG`. The naming and interface of functions is chosen so as to be familiar both to Mathematica users and users of igraph's Python and R interfaces.
 
 IGraph/M fully leverages the igraph C library's capabilities to integrate into high-level host languages: Almost all computations are interruptible in the usual manner. This feature is particularly important for a graph theory package that provides multiple algorithms with exponential or slower time complexity. Despite being implemented in a compiled language, IGraph/M uses Mathematica's built-in random number generator by default. Therefore, all its stochastic algorithms respect seeds set with `SeedRandom[]` or `BlockRandom[]` the same way as built-in functions would.
 
@@ -80,7 +80,7 @@ Load the package:
 
 ```
 In[1]:= << IGraphM`
-Out[1]= IGraph/M 0.6.2 (July 25, 2022)
+Out[1]= IGraph/M 0.6.4 (December 17, 2022)
         Evaluate IGDocumentation[] to get started.
 ```
 
