+++ IGraphM/PacletInfo.m
@@ -4,7 +4,7 @@
 
 Paclet[
     Name -> "IGraphM",
-    Version -> "0.6.4",
+    Version -> "0.6.4.1",
     MathematicaVersion -> "11.0+",
     Description -> "IGraph/M \[Dash] the igraph interface for Mathematica.",
     Creator -> "Szabolcs Horvát <szhorvat@gmail.com>",
