### 2022-12-21 285f5fb
more error checks in C++ code
