### 2022-12-21 2e0ffe9
IGBetaWeightedGabrielGraph now checks the value fo the "BetaCutoff" option
