### 2022-12-21 #4899 @szhorvat
I needed to release another version to help deal with an issue I suspect exits in the newly released Mathematica 13.2.0 on Apple Silicon.
