### 2023-01-07 #4899 @prashjha
Hi @szhorvat, can you take care of these minor typos in your JOSS draft?
1. Line 76. Reference to the section is missing in `The output is ...`.
2. Lines 78 to 97. Maybe using bullets for these examples will help. Also, Line 97 is probably incomplete. Please check.

Once you take care of these minor suggestions, I will hand your paper to EiC for final decision. Sorry for the delay.
