### 2023-01-07 #4899 @szhorvat
> Also, @szhorvat, can you please do (if not done already) a 'tagged' release of your code and archive the release using zenodo or other methods? Make sure that the zenodo archive's title matches this JOSS submission's title.

@prashjha Can you please clarify what you need archived on Zenodo? The branch that contains the paper? Or the software itself?  The Zenodo master DOI for the software is here: https://doi.org/10.5281/zenodo.1134932 The title is the software's name, not the paper's title. In order to change this, believe I need to create an entirely new release, as Zenodo does not usually allow modifying existing submissions. I wouldn't want to do that more than once, so I want to be sure I understand what you're asking.
