### 2023-01-08 #4899 @prashjha
HI @szhorvat, yes, zenodo for software. You already have the zenodo archive, so you are good. One minor thing though: change the title of the zenodo archive to match the title of your JOSS paper.

Also, please incorporate [my suggestions](https://github.com/openjournals/joss-reviews/issues/4899#issuecomment-1374289910) in the draft.

Let me know when both tasks are done. I will hand it over to the EiC for a final decision.
