### 2023-01-08 #4899 @szhorvat
Hello @prashjha,

> Also, @szhorvat, can you please do (if not done already) a 'tagged' release of your code and archive the release using zenodo or other methods? Make sure that the zenodo archive's title matches this JOSS submission's title.

This is now done. I was under the impression that Zenodo submissions were not editable, but it turns out that their _metadata_ can still be changed.

> * Line 76. Reference to the section is missing in `The output is ...`.

This was meant to be a reference to a figure, which got accidentally removed at the very beginning of the submission process. Sorry about that. It is now fixed.

> * Lines 78 to 97. Maybe using bullets for these examples will help. Also, Line 97 is probably incomplete. Please check.

Unfortunately, using bullet points would push the code into the right margin, or require a reformatting which is not helpful for readability (see screenshot below). Therefore, I would prefer to keep it as it is. Let me know if this is acceptable. If not, I'll make this change as well.
