### 2023-01-08 #4899 @danielskatz
@szhorvat - even though I've accepted this (which is in progress), I now see that the zenodo archive's metadata doesn't have any authors besides you. Can you please edit the metadata (which doesn't require a new file or change the DOI) to include all the authors of the papers?
